backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
supabase==2.3.1
gotrue>=2.9.1,<3.0.0
httpx>=0.24.1,<0.25.0
pydantic>=2.6.1,<3.0.0
python-jose[cryptography]>=3.3.0
python-dateutil>=2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
PyJWT>=2.8.0 